# openapi/pulse.yaml
openapi: 3.1.0
info:
  title: Pulse API
  version: 0.1.0
  description: |
    Pulse V2 — Personal Life OS backend.
    **Spec-first**: add every endpoint here before implementing it.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://api.pulse.app
    description: Production (Render)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/connections/request:
    post:
      tags: [Connections]
      summary: Send a connection request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_user_id]
              properties:
                to_user_id:
                  type: string
                  format: uuid
                note:
                  type: string
      responses:
        '201': { description: Sent }
        '409': { description: Already requested / connected }
...
  /api/connections/requests/{id}/accept:
    patch:
      tags: [Connections]
      summary: Accept a connection request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Accepted }
  /api/connections/requests/{id}/decline:
    patch:
      tags: [Connections]
      summary: Decline a connection request
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Declined }
...
  /api/connections/{peerId}:
    delete:
      tags: [Connections]
      summary: Remove a connection
      parameters:
        - { name: peerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }
  /api/connections/{peerId}/block:
    patch:
      tags: [Connections]
      summary: Block a user
      parameters:
        - { name: peerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Blocked }
  /api/connections/{peerId}/unblock:
    patch:
      tags: [Connections]
      summary: Unblock a user
      parameters:
        - { name: peerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Unblocked }

  # ── Teams ───────────────────────────────────────────────────────────────────
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications (unread first)
      parameters:
        - { name: limit,  in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: unread, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      type: object
                  unread_count:
                    type: integer
...
  /api/notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
  /api/notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Delete a notification
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Deleted }
  /api/notifications/push/subscribe:
    post:
      tags: [Notifications]
      summary: Register a push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, p256dh, auth]
              properties:
                endpoint: { type: string, format: uri }
                p256dh:   { type: string }
                auth:     { type: string }
      responses:
        '201': { description: Subscribed }
    delete:
      tags: [Notifications]
      summary: Unregister a push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint: { type: string }
      responses:
        '200': { description: Unsubscribed }
  /api/notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences
      responses:
        '200': { description: OK }
    put:
      tags: [Notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channels:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      in_app:   { type: boolean }
                      push:     { type: boolean }
                      telegram: { type: boolean }
                      email:    { type: boolean }
                quiet_start: { type: string, nullable: true }
                quiet_end:   { type: string, nullable: true }
      responses:
        '200': { description: OK }